feat: add initial parameter value check in Fitter
The code changes in `fit.py` add a check to ensure that the initial parameter values provided to the `Fitter` class are within the specified priors. This check is performed for each free parameter, using the corresponding prior function. If the initial value is not within the prior range, a `ValueError` is raised.