feat: add planet param validity check
This allows to check if a Planet object's default basis parameter values are valid, so we can raise an Exception or return -inf probability as required for the situation.

This acts as a last resort catch-all for any invalid values the priors may have missed.

This implementation is not final, and could feasibly be moved to the priors objects in a future refactor.